requirements.txt
django
psycopg2-binary
scrapy
requests
scrapy-user-agents
openai
python-dotenv